Rest of implementation

+++ SharedModels/WeatherModels/List.cs
@@ -14,7 +14,7 @@ namespace SharedModels.WeatherModels
         public Clouds clouds { get; set; }
         public Wind wind { get; set; }
         public int visibility { get; set; }
-        public int pop { get; set; }
+        public double pop { get; set; }
         public Sys sys { get; set; }
         public string dt_txt { get; set; }
     }
